Add custom Shubox type definitions (sort of?)
Adding a custom type definition for a project, NOT a library, is kind of
a mystery. There's no real, good, canonical approach to doing so. Like:

1. Where do you place your type definitions?
2. What do you name it?
3. What's the best way to import it?

Say I had a new type definition file I wanted to create for a new type -- `Wumpus`?
Where does that go, just for a project? How do I pull it in? How does my
project know it exists?

I sort of flailed my way around this until I got to this point where I
declared the Shubox namespace and a couple of types. I'm sure it's
probably wrong but, ¯\_(ツ)_/¯ ... it'll do for now until someone reads
this commit message (maybe?) and tells me how to do it the right way.