Refactor and allow for multiple replacements
Instead of just replacing the s3url in the final success callback, let's
allow for multiple keys in the file object to be used when interpolating
the string in the form element.

We can now use:

* 'height',
* 'width',
* 'name',
* 's3',
* 's3url',
* 'size',
* 'type',

... as replaceable "{{foo}}" placeholders in our textarea elements.

Eg the github form example:

```javascript
const githubForm = new Shubox('#shubox--textarea', {
  uuid: window.shuboxSandboxUUID,
  clickable: '#shubox--click-to-upload',
  s3urlTemplate: '![{{name}}]({{s3url}})',
  uploadingTemplate: '![Uploading {{name}}...]()'
})
```

This will say that it's uploading the "file name" while it's going to
S3, and once it's done, will replace it with the markdown image embed.