Several bug fixes
1. Instances being saved in `Shubox.instances` were being overwritten by
assigning to the same index if using several different shubox upload
targets. The fix: push onto the array, do not assign to an index more
than once.

2. Fix issue with the `s3urlTemplate` option existing in the default
option. We need to take it out as a default. If someone passes it along
as an option in their setup? Ok _then_ use it. Otherwise we were finding
it via the default hash and using it before `successTemplate`.

3. Rename the github page to /github.

4. Add /form-insert-at to the pages to show how to place the resulting
s3url in a textarea - at the cursor, append at the end, or wholly replace.

5. Check for window.console before using it.