Handle maxsize option front + back end
Missed having the maxFilesize option so added it here. In exploring how
to enforce a max filesize based on the type of account with Shubox,
noticed that it wasn't something that could be modified at runtime. It's
set at initialization.

As a result, tested with some constraints on the backend. This raised up
the issue that I was not properly handling errors returned from
`/signature`. This commit also updates the callbacks to handle an error
status and error message from that endpoint, AND delegates to the
`error` callback method - either dropzone's, shubox's, or one provided
by the user.