Add `cdn` option so final files fetched via CDN
The changes here allow for the `cdn` option. That means

```javascript
const cropped_avatar = new Shubox('#avatar-cropped', {
  key: 'bf897b63-50f2-4055-b4f6-825e613f3d3e',
  cdn: 'https://cdn.my-domain.com'
})
```

Will replace the Amazon S3 specific URL with the provided CDN http
scheme and hostname. This also tacks a cache-busting querystring to the
code that polls for generated versions/variants of an upload. Provided
the CDN caches for the unique address + querystring, instead of caching just
the address, this will work fine.