Fix bug re: callbacks not passing correct object
Come to find out that, while testing the addedfile callback, the
objects we expect to pass along to dropzone's definition of that
callback method don't end up being the right thing. Dropzone expects to
called from the context of `this` being an instance of itself. Turns out
the context was, instead, an instance of our ShuboxCallbacks class. We
needed to hold onto `this` as dropzone, and assign the `ShuboxCallbacks`
instance to a `self` object outside of the scope of the callback
definitions.

This allows us to pass along the Dropzone object, AND have access to the
ShuboxCallbacks from the context of those methods.