Smarter variant URL creation
Previously this code only handled resized images, and NOT the other file
types that can be generated.  For example - if you wanted to watch for a
new webp file, the transformCallbacks would not allow you to watch for
them.

With this update you can watch for the resized file, the new file
versions, or a combination of the two. EG:

* `'200x200'`
* `'.webp'`
* `'200x200.webp'`