Fix bug related to discerning size of a hash
`{}.length` does not work all that well. In fact it doesn't at all.
`Object.keys({})`, however - yeah that works fine.

The former is what sort of thing comes from using:

```
for (const variant of callbacks) {
```

Which was a recent change after doing some `tslint`ing. :shrug:.
Thankfully it was caught.

The latter, however, is the correct approach.

```
for (const variant of Object.keys(callbacks)) {
```