Add a message in console to help w/common localhost issue
The issue:
==========

sometimes people forget that when they're testing things out
on a local development server Shubox won't work. If you haven't set up a
"domain" and bucket for your localhost and port (likely) you'll need to
use the "Shubox Sandbox Key" that works for any running localhost instance.

That "sandbox" key is set up to upload to a temp AWS bucket that Shubox owns.

The change:
===========

This change just adds a little console.log magic in there to give a
heads up in the case the host is localhost AND the error message that we
get back from our api server says something like "Referring domain not
permitted". Easy peasy.