The best first try at this is to allow for the option to capture a photo for the current element with a string "photo" (in the future -- "video"), or via an object encapsulating the moving parts to get this to work. For example:
```
new Shubox("#webcam", {
  key: "shubox-key",
  webcam: "photo",
  success: webcamSuccessCallback;
})
```

Which will initialize an element - "#webcam" - that, when clicked, will create
a video element. Once that video element is clicked, it will capture a photo
and call the typical pipeline of callbacks.

More control can be gained for the UI with a nested hash pointing to various
elements that start and stop the parts of the camera lifecycle, like starting
and stopping the camera, and taking the photo.

```
new Shubox("#webcam", {
  key: this.shuboxKey,
  webcam: {
    type: "photo",
    startCamera: "#start_button",
    stopCamera: "#stop_button",
    startCapture: "#capture_button"
  },
  success:this._webcamSuccess
})
```

Keep in mind that the string `"photo"` infers that there are other types of
camera usage. For now there is only the ability to capture a single frame. In
the future we plan to add video (and with it, audio) capture as well.