Refactor webcam and add video recording support
Why?
====

In keeping with the theme of webcam support, I'm adding support for
recording to webm video files. This is a nice bonus feature for anyone
that wants to get "confessional" style videos added quickly from their
web-app uploaded to S3.

What?
=====

Refactor all the webcam code to split out the dom and event handling to
2 separate classes, for both the photo and video functions -- for a
total of 4 classes.

The common interface works nicely and the code is cleaned up quite a
bit.

How to test?
============

```
yarn start
open http://localhost:9001/webcam-photo
```