Add callbacks to the webcam events
Why?
====

Because sometimes you want stuff to happen after each event fires,
right? This is javascript, so ... callbacks.

What?
=====

Add some options to the webcam hash that correspond to:

* `cameraStarted`
* `cameraStopped`
* `recordingStarted`
* `recordingStopped`
* `photoTaken`

All receive the `Webcam` object as the first parameter, `photoTaken` and
`recordingStopped` receive an additional parameter that is of type
`Blob`.