Add and populate select lists IF present
Why?
====

So often people's computer setups have several different inputs and
outputs with regards to cameras, webcams, audio outputs, etc. Without
the ability to allow people to _choose_ their devices then we restrict
the ability of people to use this webcam feature.

So, we need some means to choose those devices.

What?
=====

This commit starts building out the _optional_ means to select those
devices. If the person using the JS adds a photo or video template
containing `<select>` tags with the proper classnames, then the JS lib
will take those, hook into them, and allow the changes to be reflected
in the UI/UX. Select tags can be styled and are semantically the best
method to toggle between these devices.

What selectors are looked for when we initialize things?

* `select.shubox-audioinput`
* `select.shubox-audiooutput`
* `select.shubox-videoinput`