Refactor to require providing selectors in options
Ok a few more changes. The Device selector has been pulled up to become
initialized in the webcam object. We now pass in the events handler, and
the webcam options, because ...

Instead of relying on looking for a predefined duo of selectors for the
camera and mic `<select/>`s we now require the user to pass along the
selectors to the IWebcamOptions hash. Why? Because otherwise if there
are multiple instances of a webcam capture on a page, the `<select/>`s
would act upon every one of them.

If we tie individual selectors to the instance of the Shubox/Webcam
object then we circumvent that. This refactoring now makes more sense
in terms of how things are structured.

Additionally, I moved the IWebcamOptions interface definition into the
Webcam source file itself, it shouldn't belong in the main index file.