Update s3Key details and add CHANGES.md
With the addition of string interpolation on the s3Key side of things,
it's important to surface it as an option from the README.

For further visibility of these sorts of changes, it's helpful to add
something that details all of the good new stuff when it comes to the
source. CHANGES.md will serve as the changelog/news file.