Get video recording working on Safari
This one was tough. Safari doesn't make this easy. It's very frustrating.

First of all - they put the MediaRecorder api behind a flag you have to
enable in the application's menu-bar.

`Develop -> Experimental Features -> MediaRecorder`

Second, there are portions of the common API that do not exist yet.
Namely the `MediaRecorder.isTypeSupported()` method, and recording in slices/chunks.
For the former, you just detect if the method exists. If not, use mp4.
For the latter, you have to implement an `onstop` callback to get the
contents of the final recorded stream. As things were constructed
previously, this JS had access to an array of all of the chunks at any
time - not at the point where `onstop` was triggered.

More info here: https://blog.addpipe.com/safari-technology-preview-73-adds-limited-mediastream-recorder-api-support/