Fix issue w/uploadingTemplate not replaced by successTemplate
Issue:
======

See comments @ bottom here from @germsvel :

https://github.com/thoughtbot/constable/issues/909

Why?
====

There is a bug where the uploadingTemplate isn't replaced by the
successTemplate. It's just appended instead.

What?
=====

Check where we are and whether we have to replace previous existing text
(the uploadingTemplate text) with the new text (the successTemplate).

If we're doing this, then we have to also make sure that the cursor is
placed in the right place after we change the values inside the form
elements.