Fix bug with missing portraitMode key in webcamOptions
Before this commit, if you set up a `webcamOptions` hash, and it does not
contain the `portraitMode` option, it evaluates to `undefined`. Which is falsey.

As a result, the video element in the existing logic would have its width and height
flipped. Causing a portrait mode size, for a landscape mode container.

Flipping the logic fixes this.