Consolidate transformName and transformCallbacks
Merge these two into one options: `transforms`.

This change means that we no longer rely on a transform pipeline to be
defined in `dashboard.shubox.io`. Everything is done in the `Shubox` initializer.

A `transforms` key, with a hash value containing (string) keys defining either/or
an ImageMagick geometry, or image format, will be passed along with the upload
and will kick off the work to resize and (re-)format images whatever is in the
key string.

The associated _value_ for that string key is a callback that will be run once
that file exists on s3 and is found by the Shubox library via a progressive
fallback polling of that URL's `HEAD` request.

Also:

* Pass the version of the Shubox client to the API endpoints. It may respond
  differently based on version.
* Look for proper URL's of generated/created files based on the API version
  sent _back_ in responses.
* Deprecation warnings for the old values. Will take it out in version 1.2
* Update README to reflect new configuration options